create/update some files in service laye

+++ Yoga_API/src/main/java/Refactor_Yoga/Refactor_Yoga/Service/ClientService.java
@@ -7,14 +7,19 @@ import Refactor_Yoga.Refactor_Yoga.repository.ClientRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
 public class ClientService implements BaseService<ClientDTO , Client> {
 
+    private  Client client ;
     private ClientRepository clientRepository ;
     private ClientMapper  clientMapper ;
+
+    public ClientService() {
+    }
 
     public ClientService(ClientRepository clientRepository, ClientMapper clientMapper) {
         this.clientRepository = clientRepository;
@@ -25,7 +30,7 @@ public class ClientService implements BaseService<ClientDTO , Client> {
     public ClientDTO getById(String id) {
         UUID uuid = UUID.fromString(id) ;
 
-        Client client = clientRepository.getById(uuid) ;
+         client = clientRepository.getById(uuid) ;
 
         return clientMapper.ClientToClientDTO(client) ;
     }
@@ -42,18 +47,23 @@ public class ClientService implements BaseService<ClientDTO , Client> {
     }
 
 
-    @Override
-    public void delete(String id) {
-        UUID uuid = UUID.fromString(id) ;
 
-    clientRepository.deleteById(uuid);
+    public void delete(String id) {
+
+        UUID uuid = UUID.fromString(id) ;
+        if(clientRepository.existsById(uuid))
+            clientRepository.deleteById(uuid);
+        else
+            throw new NoSuchElementException();
+
+
 
     }
 
     @Override
     public void save( ClientDTO object)
     {
-        Client client = clientMapper.ClientDTOToClient(object);
+        client = clientMapper.ClientDTOToClient(object);
         clientRepository.save(client);
     }
